[IMP] Moved manage routes to namespace app.blueprints.pages
