+++ source/app/blueprints/pages/manage/manage_assets_type_routes.py
