+++ source/app/blueprints/pages/manage/manage_attributes_routes.py
