+++ source/app/blueprints/pages/manage/manage_case_state.py
