+++ source/app/blueprints/pages/manage/manage_case_templates_routes.py
