+++ source/app/blueprints/pages/manage/manage_ioc_types_routes.py
