+++ source/app/blueprints/pages/manage/manage_objects_routes.py
