+++ source/app/blueprints/pages/manage/templates/manage_attributes.html
