+++ source/app/blueprints/pages/manage/templates/manage_case_templates.html
