+++ source/app/blueprints/pages/manage/templates/manage_cases.html
