+++ source/app/blueprints/pages/manage/templates/manage_customer_view.html
