+++ source/app/blueprints/pages/manage/templates/manage_objects.html
