+++ source/app/blueprints/pages/manage/templates/modal_add_asset_type.html
