+++ source/app/blueprints/pages/manage/templates/modal_add_attribute.html
