+++ source/app/blueprints/pages/manage/templates/modal_add_case.html
