+++ source/app/blueprints/pages/manage/templates/modal_add_ioc_type.html
