+++ source/app/blueprints/pages/manage/templates/modal_case_info_from_case.html
