+++ source/app/blueprints/pages/manage/templates/modal_case_state.html
