+++ source/app/blueprints/pages/manage/templates/modal_module_info.html
