+++ source/app/blueprints/pages/manage/templates/modal_server_updates.html
