+++ source/app/blueprints/pages/manage/templates/modal_upload_case_template.html
