+++ source/app/views.py
@@ -34,22 +34,22 @@ from app.blueprints.pages.datastore.datastore_routes import datastore_blueprint
 from app.blueprints.pages.demo_landing.demo_landing import demo_blueprint
 from app.blueprints.dim_tasks.dim_tasks import dim_tasks_blueprint
 from app.blueprints.login.login_routes import login_blueprint
-from app.blueprints.manage.manage_access_control import manage_ac_blueprint
-from app.blueprints.manage.manage_assets_type_routes import manage_assets_type_blueprint
-from app.blueprints.manage.manage_attributes_routes import manage_attributes_blueprint
-from app.blueprints.manage.manage_case_classifications import manage_case_classification_blueprint
-from app.blueprints.manage.manage_case_state import manage_case_state_blueprint
-from app.blueprints.manage.manage_evidence_types_route import manage_evidence_types_blueprint
-from app.blueprints.manage.manage_cases_routes import manage_cases_blueprint
-from app.blueprints.manage.manage_customers_routes import manage_customers_blueprint
-from app.blueprints.manage.manage_groups import manage_groups_blueprint
-from app.blueprints.manage.manage_ioc_types_routes import manage_ioc_type_blueprint
-from app.blueprints.manage.manage_modules_routes import manage_modules_blueprint
-from app.blueprints.manage.manage_objects_routes import manage_objects_blueprint
-from app.blueprints.manage.manage_srv_settings_routes import manage_srv_settings_blueprint
-from app.blueprints.manage.manage_templates_routes import manage_templates_blueprint
-from app.blueprints.manage.manage_case_templates_routes import manage_case_templates_blueprint
-from app.blueprints.manage.manage_users import manage_users_blueprint
+from app.blueprints.pages.manage.manage_access_control import manage_ac_blueprint
+from app.blueprints.pages.manage.manage_assets_type_routes import manage_assets_type_blueprint
+from app.blueprints.pages.manage.manage_attributes_routes import manage_attributes_blueprint
+from app.blueprints.pages.manage.manage_case_classification_routes import manage_case_classification_blueprint
+from app.blueprints.pages.manage.manage_case_state import manage_case_state_blueprint
+from app.blueprints.pages.manage.manage_evidence_types_route import manage_evidence_types_blueprint
+from app.blueprints.pages.manage.manage_cases_routes import manage_cases_blueprint
+from app.blueprints.pages.manage.manage_customers_routes import manage_customers_blueprint
+from app.blueprints.pages.manage.manage_groups_routes import manage_groups_blueprint
+from app.blueprints.pages.manage.manage_ioc_types_routes import manage_ioc_type_blueprint
+from app.blueprints.pages.manage.manage_modules_routes import manage_modules_blueprint
+from app.blueprints.pages.manage.manage_objects_routes import manage_objects_blueprint
+from app.blueprints.pages.manage.manage_srv_settings_routes import manage_srv_settings_blueprint
+from app.blueprints.pages.manage.manage_templates_routes import manage_templates_blueprint
+from app.blueprints.pages.manage.manage_case_templates_routes import manage_case_templates_blueprint
+from app.blueprints.pages.manage.manage_users import manage_users_blueprint
 from app.blueprints.pages.overview.overview_routes import overview_blueprint
 from app.blueprints.pages.profile.profile_routes import profile_blueprint
 from app.blueprints.pages.search.search_routes import search_blueprint
